requirements.txt
python-telegram-bot 